backend/pyproject.toml
[project]
name = "group-todo"
version = "0.1.0"
description = "Collaborative todo list"
authors = [
    {name = "Chris La Valle", email = "codex@openai.com"},
    {name = "Codex", email = "codex@openai.com"},
]
license = "MIT"
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "Django>=5,<6",
    "djangorestframework",
    "channels",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
